refactor(agents): split subagent-registry-run-manager.ts into owner modules
